requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.8
starlette~=0.45.3
python-dotenv~=1.0.1
mysql-connector-python~=9.2.0
pydantic~=2.10.6
pandas~=2.2.3
pymongo~=4.11.1
pydantic_core~=2.27.2